+++ lsp-server/diplomat/include/diplomat_lsp.hpp
@@ -64,6 +64,7 @@ class DiplomatLSP : public slsp::BaseLSP
         json _h_get_module_bbox(json params);
         void _h_set_module_top(json params);
         void _h_ignore(json params);
+        void _h_add_to_include(json params);
         void _h_force_clear_index(json params);
 
         json _h_resolve_hier_path(json params);